requirements.txt
Django==1.11.10
djangorestframework==3.7.7
psycopg2==2.7.4
pytz==2018.3
